+++ packages/ag-grid-community/src/selection/selectionService.ts
@@ -6,14 +6,12 @@ import type { SelectionEventSourceType } from '../events';
 import { isSelectionUIEvent } from '../events';
 import {
     _getGroupSelectsDescendants,
-    _getIsRowSelectable,
     _getRowSelectionMode,
     _isClientSideRowModel,
     _isMultiRowSelection,
     _isUsingNewRowSelectionAPI,
 } from '../gridOptionsUtils';
 import type { IClientSideRowModel } from '../interfaces/iClientSideRowModel';
-import type { IRowModel } from '../interfaces/iRowModel';
 import type { ISelectionService, ISetNodesSelectedParams } from '../interfaces/iSelectionService';
 import type { ServerSideRowGroupSelectionState, ServerSideRowSelectionState } from '../interfaces/selectionState';
 import type { PageBoundsService } from '../pagination/pageBoundsService';
@@ -27,11 +25,10 @@ import { RowRangeSelectionContext } from './rowRangeSelectionContext';
 export class SelectionService extends BaseSelectionService implements NamedBean, ISelectionService {
     beanName = 'selectionService' as const;
 
-    private rowModel: IRowModel;
     private pageBoundsService: PageBoundsService;
 
-    public wireBeans(beans: BeanCollection): void {
-        this.rowModel = beans.rowModel;
+    public override wireBeans(beans: BeanCollection): void {
+        super.wireBeans(beans);
         this.pageBoundsService = beans.pageBoundsService;
     }
 
@@ -41,7 +38,8 @@ export class SelectionService extends BaseSelectionService implements NamedBean,
     private groupSelectsChildren: boolean;
     private rowSelectionMode?: RowSelectionMode = undefined;
 
-    public postConstruct(): void {
+    public override postConstruct(): void {
+        super.postConstruct();
         const { gos, rowModel, onRowSelected } = this;
         this.selectionCtx.init(rowModel);
         this.rowSelectionMode = _getRowSelectionMode(gos);
@@ -56,8 +54,6 @@ export class SelectionService extends BaseSelectionService implements NamedBean,
                 this.deselectAllRowNodes({ source: 'api' });
             }
         });
-
-        this.addManagedPropertyListener('isRowSelectable', () => this.updateSelectable(false));
 
         this.addManagedEventListeners({ rowSelected: onRowSelected.bind(this) });
     }
@@ -275,7 +271,10 @@ export class SelectionService extends BaseSelectionService implements NamedBean,
     }
 
     // should only be called if groupSelectsChildren=true
-    public updateGroupsFromChildrenSelections(source: SelectionEventSourceType, changedPath?: ChangedPath): boolean {
+    public override updateGroupsFromChildrenSelections(
+        source: SelectionEventSourceType,
+        changedPath?: ChangedPath
+    ): boolean {
         // we only do this when group selection state depends on selected children
         if (!this.groupSelectsChildren) {
             return false;
@@ -663,73 +662,6 @@ export class SelectionService extends BaseSelectionService implements NamedBean,
             throw new Error(
                 `selectAll only available when rowModelType='clientSide', ie not ${this.rowModel.getType()}`
             );
-        }
-    }
-
-    /**
-     * Updates the selectable state for a node by invoking isRowSelectable callback.
-     * If the node is not selectable, it will be deselected.
-     *
-     * Callers:
-     *  - property isRowSelectable changed
-     *  - after grouping / treeData
-     */
-    public updateSelectable(skipLeafNodes: boolean) {
-        const { gos } = this;
-
-        const isRowSelecting = _getRowSelectionMode(gos) !== undefined;
-        const isRowSelectable = _getIsRowSelectable(gos);
-
-        if (!isRowSelecting || !isRowSelectable) {
-            return;
-        }
-
-        const isGroupSelectsChildren = _getGroupSelectsDescendants(gos);
-        const isCsrmGroupSelectsChildren = _isClientSideRowModel(gos) && isGroupSelectsChildren;
-
-        const nodesToDeselect: RowNode[] = [];
-
-        const nodeCallback = (node: RowNode) => {
-            if (skipLeafNodes && !node.group) {
-                return;
-            }
-
-            // Only in the CSRM, we allow group node selection if a child has a selectable=true when using groupSelectsChildren
-            if (isCsrmGroupSelectsChildren && node.group) {
-                const hasSelectableChild = node.childrenAfterGroup!.some((rowNode) => rowNode.selectable === true);
-                node.setRowSelectable(hasSelectableChild, true);
-                return;
-            }
-
-            const rowSelectable = isRowSelectable?.(node) ?? true;
-            node.setRowSelectable(rowSelectable, true);
-
-            if (!rowSelectable && node.isSelected()) {
-                nodesToDeselect.push(node);
-            }
-        };
-
-        // Needs to be depth first in this case, so that parents can be updated based on child.
-        if (isCsrmGroupSelectsChildren) {
-            const csrm = this.rowModel as IClientSideRowModel;
-            const changedPath = new ChangedPath(false, csrm.getRootNode());
-            changedPath.forEachChangedNodeDepthFirst(nodeCallback, true, true);
-        } else {
-            // Normal case, update all rows
-            this.rowModel.forEachNode(nodeCallback);
-        }
-
-        if (nodesToDeselect.length) {
-            this.setNodesSelected({
-                nodes: nodesToDeselect,
-                newValue: false,
-                source: 'selectableChanged',
-            });
-        }
-
-        // if csrm and group selects children, update the groups after deselecting leaf nodes.
-        if (isCsrmGroupSelectsChildren) {
-            this.updateGroupsFromChildrenSelections('selectableChanged');
         }
     }
 }
