+++ packages/ag-grid-enterprise/src/rowGrouping/groupStage/groupStrategy/groupStrategy.ts
@@ -4,7 +4,6 @@ import type {
     ChangedPath,
     ColumnModel,
     FuncColsService,
-    IRowNodeStage,
     ISelectionService,
     IShowRowGroupColsService,
     InitialGroupOrderComparatorParams,
@@ -428,7 +427,7 @@ export class GroupStrategy extends BeanStub {
         }
 
         // groups are about to get disposed, so need to deselect any that are selected
-        this.selectionService?.filterFromSelection((node: RowNode) => node && !node.group);
+        this.selectionService?.filterFromSelection?.((node: RowNode) => node && !node.group);
 
         const { groupedCols } = details;
         const rootNode: GroupRow = details.rootNode;
